move seed node out
