+++ python/dgl/graphbolt/node_unified_data_struct.py
@@ -6,21 +6,13 @@ from typing import Dict, Union
 
 import torch
 
-from ..unified_data_struct import UnifiedDataStruct
+from .unified_data_struct import UnifiedDataStruct
 
 
 @dataclass
 class NodeUnifiedDataStruct(UnifiedDataStruct):
     r"""A subclass of 'UnifiedDataStruct', specialized for handling node level
     tasks."""
-
-    seed_node: Union[torch.Tensor, Dict[str, torch.Tensor]]
-    """
-    Representation of seed nodes in the graph.
-    - If `seed_node` is a tensor: It indicates the graph is homogeneous.
-    - If `seed_node` is a dictionary: The keys should be node type and the
-      value should be corresponding heterogeneous node ids.
-    """
 
     label: Union[torch.Tensor, Dict[str, torch.Tensor]] = None
     """
